pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["VladT7 <vladt13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
slack-sdk = "^3.33.3"
python-dotenv = "^1.0.1"
flask = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
